requirements.txt
playwright
beautifulsoup4
requests
sentence-transformers
faiss-cpu
pandas
openpyxl
apscheduler
python-dotenv